+++ src/App.stories.tsx
@@ -1,9 +1,9 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import App from "../App";
+import type { Meta, StoryObj } from '@storybook/react';
+import App from './App';
 
 const meta = {
-  title: "App",
-  component: App,
+  title: 'App',
+  component: App
 } satisfies Meta<typeof App>;
 
 export default meta;
